+++ Sources/TuistGraph/Models/ProjectOption.swift
@@ -5,11 +5,16 @@ public enum ProjectOption: Codable {
     /// Text settings to override user ones for current project
     case textSettings(TextSettings)
 
+    /// Disable the synthesized resource accessors generation
+    case disableSynthesizedResourceAccessors
+
     /// Option name
     public var name: String {
         switch self {
         case .textSettings:
             return "textSettings"
+        case .disableSynthesizedResourceAccessors:
+            return "disableSynthesizedResourceAccessors"
         }
     }
 }
@@ -22,8 +27,14 @@ extension Array where Element == ProjectOption {
             switch $0 {
             case let .textSettings(textSettings):
                 return textSettings
+            case .disableSynthesizedResourceAccessors:
+                return nil
             }
         }.first
+    }
+
+    public var disableSynthesizedResourceAccessors: Bool {
+        return self.contains(.disableSynthesizedResourceAccessors)
     }
 }
 
@@ -36,8 +47,10 @@ extension ProjectOption: Hashable {
 
     public static func == (lhs: ProjectOption, rhs: ProjectOption) -> Bool {
         switch (lhs, rhs) {
-        case (.textSettings, .textSettings):
+        case (.textSettings, .textSettings), (.disableSynthesizedResourceAccessors, .disableSynthesizedResourceAccessors):
             return true
+        case (.textSettings, .disableSynthesizedResourceAccessors), (.disableSynthesizedResourceAccessors, .textSettings):
+            return false
         }
     }
 }
@@ -47,12 +60,15 @@ extension ProjectOption: Hashable {
 extension ProjectOption {
     internal enum CodingKeys: String, CodingKey {
         case textSettings
+        case disableSynthesizedResourceAccessors
     }
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
 
-        if container.allKeys.contains(.textSettings), try container.decodeNil(forKey: .textSettings) == false {
+        if container.contains(.disableSynthesizedResourceAccessors) {
+            self = .disableSynthesizedResourceAccessors
+        } else if container.allKeys.contains(.textSettings), try container.decodeNil(forKey: .textSettings) == false {
             var associatedValues = try container.nestedUnkeyedContainer(forKey: .textSettings)
             let textSettings = try associatedValues.decode(TextSettings.self)
             self = .textSettings(textSettings)
@@ -68,6 +84,8 @@ extension ProjectOption {
         case let .textSettings(textSettings):
             var associatedValues = container.nestedUnkeyedContainer(forKey: .textSettings)
             try associatedValues.encode(textSettings)
+        case .disableSynthesizedResourceAccessors:
+            try container.encode(true, forKey: .disableSynthesizedResourceAccessors)
         }
     }
 }